requirements.txt
requests>=2.31.0
redis>=5.0.0
openai>=1.0.0
contextual-client>=0.1.0
python-dotenv>=1.0.0
